requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
selenium==4.26.1
ebooklib==0.18
Pillow==11.0.0
lxml==5.3.0